Refs #46: Rename `SynchroniseWithShopify` concern to simply `Synchronises`, and add a couple of helper methods for deciding whether to synchronise models.

+++ app/models/disco_app/concerns/synchronises_with_shopify.rb
@@ -1,10 +1,16 @@
-module DiscoApp::Concerns::SynchronisesWithShopify
+module DiscoApp::Concerns::Synchronises
   extend ActiveSupport::Concern
 
   class_methods do
 
-    def synchronise_from_shopify(shop, data)
+    def should_synchronise?(shop, data)
+      true
+    end
+
+    def synchronise(shop, data)
       data = data.with_indifferent_access
+
+      return unless should_synchronise?(shop, data)
 
       instance = self.find_or_create_by!(id: data[:id]) do |instance|
         instance.shop = shop
@@ -16,8 +22,15 @@ module DiscoApp::Concerns::SynchronisesWithShopify
       instance
     end
 
-    def synchronise_deletion_from_shopify(shop, data)
+    def should_synchronise_deletion?(shop, data)
+      true
+    end
+
+    def synchronise_deletion(shop, data)
       data = data.with_indifferent_access
+
+      return unless should_synchronise_deletion?(shop, data)
+
       self.destroy_all(shop: shop, id: data[:id])
     end
